+++ docs/examples/my-fuel-dapp/frontend/src/App.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import { Wallet } from "fuels";
 import "./App.css";
 // Import the contract factory from the generated folder
-// from the typechain command
-import { ContractsAbi__factory } from "./contracts";
+// from the fuelchain command
+import { ContractAbi__factory } from "./contracts";
 
 // Se the Wallet Secret used on the chainConfig.json
 // this enables us to have a account with initial balance
@@ -11,14 +11,16 @@ const WALLET_SECRET =
   "0xa449b1ffee0e2205fa924c6740cc48b3b473aa28587df6dab12abc245d1f5298";
 // The address of the contract deployed to our local node
 // the contract id is output right after the forc deploy command
+// Ex.: Contract id: 0xa326e3472fd4abc417ba43e369f59ea44f8325d42ba6cf71ec4b58123fd8668a
+// const CONTRACT_ID = "0xa326e3472fd4abc417ba43e369f59ea44f8325d42ba6cf71ec4b58123fd8668a"
 const CONTRACT_ID =
-  "0xa326e3472fd4abc417ba43e369f59ea44f8325d42ba6cf71ec4b58123fd8668a";
+  "0xc6ff6938a90e462736aaedf19283c782a25fabf89aa496a707de8b18cf66ceb3";
 // Create a Wallet from given secretKey in this case
 // The one we configured at the chainConfig.json
 const wallet = new Wallet(WALLET_SECRET);
 // Connects out Contract instance to the deployed contract
 // address using the given wallet.
-const contract = ContractsAbi__factory.connect(CONTRACT_ID, wallet);
+const contract = ContractAbi__factory.connect(CONTRACT_ID, wallet);
 
 function App() {
   const [counter, setCounter] = useState(0);