+++ packages/typechain-target-fuels/src/codegen/index.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-use-before-define */
-import type { CodegenConfig } from 'typechain';
+import type { CodegenConfig } from 'fuelchain';
 
 import { FACTORY_POSTFIX } from '../common';
 import type { Contract, RawAbiDefinition } from '../parser/abiParser';
@@ -103,10 +103,10 @@ function codegenCommonContractFactory(
   const body = `
     static readonly abi = _abi;
     static createInterface(): ${contract.name}Interface {
-      return new Interface(_abi) as ${contract.name}Interface;
+      return new Interface(_abi) as unknown as ${contract.name}Interface;
     }
     static connect(id: string | AbstractAddress, walletOrProvider: Wallet | Provider): ${contract.name} {
-      return new Contract(id, _abi, walletOrProvider) as ${contract.name};
+      return new Contract(id, _abi, walletOrProvider) as unknown as ${contract.name};
     }
   `.trim();