+++ packages/wallet/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 0.14.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fuel-ts/abi-coder@0.14.0
+  - @fuel-ts/constants@0.14.0
+  - @fuel-ts/hasher@0.14.0
+  - @fuel-ts/hdwallet@0.14.0
+  - @fuel-ts/interfaces@0.14.0
+  - @fuel-ts/math@0.14.0
+  - @fuel-ts/mnemonic@0.14.0
+  - @fuel-ts/providers@0.14.0
+  - @fuel-ts/signer@0.14.0
+  - @fuel-ts/transactions@0.14.0
+
 ## 0.13.0
 
 ### Minor Changes